[update] Improve wallet and payments in Database

+++ app/src/main/java/co/domix/android/login/interactor/SignupInteractorImpl.java
@@ -33,7 +33,7 @@ public class SignupInteractorImpl implements SignupInteractor {
                        Activity activity, FirebaseAuth firebaseAuth) {
         String codeCountry = "";
         try {
-            geocodeMatches = new Geocoder(activity).getFromLocation(Double.valueOf(latitude), Double.valueOf(longitude), 1);
+            geocodeMatches = new Geocoder(activity).getFromLocation(Float.valueOf(latitude), Float.valueOf(longitude), 1);
         } catch (IOException e) {
             e.printStackTrace();
         }
